+++ R/utilities.R
@@ -29,6 +29,7 @@
     jmespath(db, path) |>
         parse_json(..., simplifyVector = simplifyVector)
 }
+
 .onLoad <-
     function(libname, pkgname)
 {
@@ -43,4 +44,23 @@
             }
         }
     }
+}
+
+.wrap <-
+    function(...)
+{
+    x <- paste0(...)
+    paste(strwrap(x), collapse = "\n")
+}
+
+.onAttach <-
+    function(libname, pkgname)
+{
+    packageStartupMessage(.wrap(
+        "The interface to CELLxGENE has changed; versions of ",
+        "cellxgenedp prior to 1.4.1 / 1.5.2 will cease to work when ",
+        "CELLxGENE removes the previous interface. See the section ",
+        "'API changes' of the 'Discover and download datasets...' ",
+        "vignette for important details."
+    ))
 }